+++ bgw-gui/src/commonMain/kotlin/PropData.kt
@@ -61,6 +61,7 @@ internal class AppData : Data() {
   var action: ActionProp = ActionProp.DEFAULT
   var fadeTime: Int = 0
   var blurRadius: Double = 0.0
+  var forcedByAnimation: Boolean = false
 }
 
 @Serializable
@@ -81,6 +82,7 @@ internal abstract class AnimationData : Data() {
   var onFinished: ((EventData) -> Unit)? = null
   var animationType: String = ""
   var isStop: Boolean = false
+  var interpolation: String = "linear"
 }
 
 @Serializable
@@ -178,7 +180,9 @@ internal class CheckBoxData : LabeledUIComponentData() {
 internal class ComboBoxData : UIComponentData() {
   var prompt: String = ""
   var items: List<Pair<Int, String>> = emptyList()
+  var itemVisuals: List<VisualData> = emptyList()
   var selectedItem: Pair<Int, String>? = null
+  var disallowUnselect: Boolean = false
 }
 
 @Serializable internal class LabelData : LabeledUIComponentData() {}
@@ -205,7 +209,7 @@ internal class ColorPickerData : UIComponentData() {
 @Serializable
 internal class ProgressBarData : UIComponentData() {
   var progress: Double = 0.0
-  var barColor: String = "#000000"
+  var barVisual: VisualData? = null
 }
 
 @Serializable internal class TextAreaData : TextInputUIComponentData() {}