+++ bgw-gui/src/jsMain/kotlin/tools/aqua/bgw/elements/uicomponents/ToggleButton.kt
@@ -45,7 +45,7 @@ internal fun PropertiesBuilder.cssBuilderIntern(componentViewData: ToggleButtonD
   cssBuilder(componentViewData)
   alignItems = AlignItems.center
   justifyItems = JustifyItems.flexStart
-  gap = 10.em
+  gap = 10.bgw
   pointerEvents = if (componentViewData.isDisabled) None.none else PointerEvents.all
 }
 
@@ -83,9 +83,9 @@ internal val ToggleButton =
           name = props.data.group
 
           css {
-            width = 20.em
-            height = 20.em
-            maxWidth = 20.em
+            width = 20.bgw
+            height = 20.bgw
+            maxWidth = 20.bgw
             zIndex = integer(1)
           }
           onChange = {