+++ bgw-gui/src/jsMain/kotlin/tools/aqua/bgw/event/JCEFEventDispatcher.kt
@@ -21,59 +21,34 @@ import data.event.AnimationFinishedEventData
 import data.event.EventData
 import data.event.KeyEventData
 import jsonMapper
-import kotlinx.browser.window
 import kotlinx.serialization.encodeToString
-import org.w3c.dom.WebSocket
 import tools.aqua.bgw.webSocket
 
 internal object JCEFEventDispatcher : EventDispatcher {
-  init {
-    initialize()
-  }
-
   fun dispatchGlobalEvent(event: KeyEventData) {
-    val json = jsonMapper.encodeToString(event)
     try {
-      window.asDynamic().bgwSceneQuery(Base64.encode(json))
+      val json = jsonMapper.encodeToString(event)
+      webSocket?.send("bgwGlobalQuery|$json")
     } catch (e: Throwable) {
       println("Error while dispatching event: $e")
     }
   }
 
   override fun dispatchEvent(event: AnimationFinishedEventData) {
-    val json = jsonMapper.encodeToString(event)
     try {
-      window.asDynamic().bgwAnimationQuery(Base64.encode(json))
+      val json = jsonMapper.encodeToString(event)
+      webSocket?.send("bgwAnimationQuery|$json")
     } catch (e: Throwable) {
       println("Error while dispatching event: $e")
     }
   }
 
   override fun dispatchEvent(event: EventData) {
-    val json = jsonMapper.encodeToString(event)
-    // Check WebSocket state before sending
-    if (webSocket != null && webSocket?.readyState == WebSocket.OPEN) {
-      try {
-        webSocket?.send(json)
-      } catch (e: Throwable) {
-        println("WebSocket send error: $e")
-      }
-    } else {
-      println("WebSocket is not open, cannot send event: $json")
-    }
     try {
-      window.asDynamic().bgwQuery(Base64.encode(json))
+      val json = jsonMapper.encodeToString(event)
+      webSocket?.send("bgwQuery|$json")
     } catch (e: Throwable) {
       println("Error while dispatching event: $e")
     }
-  }
-
-  private fun initialize() {
-    js(
-        "window.bgwQuery = function(request) { if(window.cefQuery) { window.cefQuery({request: request, persistent: false, onSuccess: function (response) {}, onFailure: function (error_code, error_message) {}}); console.log(Date.now()); } }")
-    js(
-        "window.bgwAnimationQuery = function(request) { if(window.cefAnimationQuery) window.cefAnimationQuery({request: request, persistent: false, onSuccess: function (response) {}, onFailure: function (error_code, error_message) {}}) }")
-    js(
-        "window.bgwSceneQuery = function(request) { if(window.cefSceneQuery) window.cefSceneQuery({request: request, persistent: false, onSuccess: function (response) {}, onFailure: function (error_code, error_message) {}}) }")
   }
 }