+++ bgw-gui/src/jvmMain/kotlin/tools/aqua/bgw/mapper/AnimationMapper.kt
@@ -74,17 +74,20 @@ internal object AnimationMapper {
             fromOpacity = animation.fromOpacity
             toOpacity = animation.toOpacity
             animationType = "fade"
+            interpolation = animation.interpolation.name.lowercase()
           }
       is MovementAnimation<*> ->
           (mapSpecific(animation) as MovementAnimationData).apply {
             byX = (animation.toX - animation.fromX).toInt()
             byY = (animation.toY - animation.fromY).toInt()
             animationType = "move"
+            interpolation = animation.interpolation.name.lowercase()
           }
       is RotationAnimation<*> ->
           (mapSpecific(animation) as RotationAnimationData).apply {
             byAngle = animation.toAngle - animation.fromAngle
             animationType = "rotate"
+            interpolation = animation.interpolation.name.lowercase()
           }
       is ScaleAnimation<*> ->
           (mapSpecific(animation) as ScaleAnimationData).apply {
@@ -93,6 +96,7 @@ internal object AnimationMapper {
             toScaleX = animation.toScaleX
             toScaleY = animation.toScaleY
             animationType = "scale"
+            interpolation = animation.interpolation.name.lowercase()
           }
       is FlipAnimation<*> ->
           (mapSpecific(animation) as FlipAnimationData).apply {